thin out variables
